refactor: organized code structure into packages
