+++ TAAM/app/src/main/java/com/example/taam/MainActivity.java
@@ -8,7 +8,10 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-import com.example.taam.home.UserHomeFragment;
+import com.example.taam.database.DataModel;
+import com.example.taam.database.Item;
+import com.example.taam.ui.home.UserHomeFragment;
+import com.example.taam.ui.FragmentLoader;
 
 public class MainActivity extends AppCompatActivity {
