+++ TAAM/app/src/main/java/com/example/taam/database/DataModel.java
@@ -1,8 +1,6 @@
-package com.example.taam;
+package com.example.taam.database;
 
-import android.content.Context;
 import android.util.Log;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
