+++ TAAM/app/src/main/java/com/example/taam/database/Item.java
@@ -1,4 +1,4 @@
-package com.example.taam;
+package com.example.taam.database;
 
 import android.util.Log;
