+++ TAAM/app/src/main/res/layout/fragment_admin_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.UserHomeFragment">
+    tools:context=".ui.home.UserHomeFragment">
 
     <Button
         android:id="@+id/view_button"
